src/requirements.txt
aiohttp==3.8.3
aiohttp_remotes==1.2.0
coloredlogs==15.0.1
asyncpg==0.29.0
pillow==10.3.0